refactor(store-25): '요일' enum을 공통화 하여 반영 사항 리팩토링
- 조회 응답 시 id를 반환하지 않아 이도 같이 추가
